requirements.txt
Django==5.1.1
channels==4.0.0
channels-redis==4.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.9 
pillow==11.3.0
python-dotenv==1.0.1
django-filter==24.2
daphne==4.1.2
gunicorn==22.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
django-cors-headers==4.3.1
drf-yasg